Extract utilities, observers and uniqueness apis to their own js files.
